+++ tests/morphy_test.py
@@ -7,7 +7,7 @@ from wn.morphy import Morphy
 
 @pytest.mark.usefixtures('mini_db')
 def test_morphy():
-    w = wn.Wordnet('test-en:1', lemmatizer_class=Morphy)
+    w = wn.Wordnet('test-en:1', lemmatizer=Morphy)
     m = w.lemmatizer
     assert list(m('examples', 'n')) == ['example']
     assert list(m('examples', 'v')) == []